Revert "Back-end code for tracking progress"
